backend/requirements.txt
alembic==1.15.1
asgiref==3.8.1
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.2
docopt==0.6.2
Faker==37.1.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pipreqs==0.4.13
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.5
python-dotenv==1.0.1
requests==2.32.4
SQLAlchemy==2.0.39
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
yarg==0.1.10
